# src/aarkib/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aarkib Media Server REST API",
    "version": "0.1.0",
    "description": "Unified REST API for Aarkib self-hosted media server. Powers native mobile and TV clients, e-readers, and web applications across books, comics, audiobooks, music, podcasts, movies, and TV shows."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Aarkib server instance"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "BasicAuth": []
    }
  ],
  "paths": {
...
    "/api/media/{id}": {
      "get": {
        "summary": "Get full media item details",
        "tags": ["Media"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Detailed media item object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItemDetail"
                }
              }
            }
          },
          "404": { "description": "Media item not found" }
        }
      }
    },
...
    "/api/tags": {
      "get": {
        "summary": "List tags / genres",
        "tags": ["Taxonomies"],
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": { "description": "Paginated list of tags" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "DeviceToken issued via /api/auth/login or /api/tokens"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Standard HTTP Basic Auth using username and password"
      }
    },
    "schemas": {
...
      "MediaItemSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "media_type": { "type": "string", "example": "movie" },
          "creators": { "type": "array", "items": { "type": "string" } },
          "creators_display": { "type": "string" },
          "file_format": { "type": "string", "example": "mp4" },
          "file_size": { "type": "integer" },
          "cover_url": { "type": "string" },
          "player_url": { "type": "string" },
          "duration": { "type": "number", "nullable": true },
          "season": { "type": "integer", "nullable": true },
          "episode": { "type": "integer", "nullable": true },
          "series_index": { "type": "number", "nullable": true },
          "collection": { "type": "string", "nullable": true },
          "progress": {
            "type": "object",
            "properties": {
              "percentage": { "type": "number" },
              "location": { "type": "string" },
              "completed": { "type": "boolean" }
            }
          }
        }
      },
      "MediaItemDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/MediaItemSummary" },
          {
            "type": "object",
            "properties": {
              "description": { "type": "string", "nullable": true },
              "publisher": { "type": "string", "nullable": true },
              "language": { "type": "string", "nullable": true },
              "isbn": { "type": "string", "nullable": true },
              "resolution_width": { "type": "integer", "nullable": true },
              "resolution_height": { "type": "integer", "nullable": true },
              "codec": { "type": "string", "nullable": true },
              "download_url": { "type": "string" },
              "file_url": { "type": "string" }
            }
          }
        ]
      },
...
    }
  }
}